+++ R/RunMAPOneSample.R
@@ -3,6 +3,10 @@
 #' @inheritParams MAPAssignActivityInternal
 #'
 #' @param output.dir Directory path to save the output file.
+#' 
+#' @param drop.low.mut.samples Whether to exclude low mutation samples from
+#' the analysis. If \code{TRUE(default)}, samples with SBS total mutations less
+#' than 100, DBS or ID total mutations less than 25 will be dropped.
 #'
 #' @importFrom utils write.csv
 #'
@@ -19,7 +23,14 @@ RunMAPOnOneSample <-
            progress.monitor        = NULL,
            seed                    = NULL,
            max.subsets             = 1000,
-           use.sparse.assign       = FALSE) {
+           use.sparse.assign       = FALSE,
+           drop.low.mut.samples    = TRUE) {
+    
+    if (drop.low.mut.samples) {
+      spect <- DropLowMutationSamples(spect)
+    } else {
+      spect <- spect
+    }
     
     if (!dir.exists(output.dir)) {
       dir.create(output.dir, recursive = TRUE)
@@ -37,7 +48,8 @@ RunMAPOnOneSample <-
       progress.monitor        = progress.monitor,
       seed                    = seed,
       max.subsets             = max.subsets,
-      use.sparse.assign       = use.sparse.assign)
+      use.sparse.assign       = use.sparse.assign,
+      drop.low.mut.samples    = drop.low.mut.samples)
     
     if (!is.null(retval$error.messages)) {
       return(retval)