+++ tests/testthat/test_DropLowMutationSamples.R
@@ -1,4 +1,6 @@
 test_that("DropLowMutationSamples for ID spectra", {
+  skip_if_not(Sys.getenv("MSIGACT_TEST_LENGTH") == "long")
+  
   spectra <- PCAWG7::spectra$PCAWG$ID
   catalog.list <- PCAWG7::SplitPCAWGMatrixByTumorType(spectra)
   lung.catalogs <- catalog.list$`Lung-AdenoCA`
@@ -13,7 +15,7 @@ test_that("DropLowMutationSamples for ID spectra", {
   sigs <- ID.sigs[, names(sigs.prop), drop = FALSE]
   my.opts <- DefaultManyOpts()
   
-  output.dir <- file.path(tempdir(), paste0("test", 1))
+  output.dir <- file.path(tempdir(), paste0("test", 1:3))
   retval1 <- expect_message(
     MAPAssignActivity(
     spectra                 = catalog,
@@ -39,6 +41,38 @@ test_that("DropLowMutationSamples for ID spectra", {
     seed                    = 2351,
     use.sparse.assign       = TRUE),
     "Samples with total mutations less than 25 were excluded in the analysis for ID")
+  
+  
+  retval3 <- MAPAssignActivity(
+    spectra                 = catalog,
+    sigs                    = sigs,
+    sigs.presence.prop      = sigs.prop,
+    output.dir              = output.dir[2],
+    max.level               = ncol(sigs) - 1,
+    p.thresh                = 0.01,
+    m.opts                  = my.opts,
+    num.parallel.samples    = 1,
+    mc.cores.per.sample     = 30,
+    seed                    = 2351,
+    drop.low.mut.samples    = FALSE)
+  
+  expect_equal(nrow(retval3$proposed.assignment), 3)
+  
+  retval4 <- MAPAssignActivity(
+    spectra                 = catalog,
+    sigs                    = sigs,
+    sigs.presence.prop      = sigs.prop,
+    output.dir              = output.dir[3],
+    max.level               = ncol(sigs) - 1,
+    p.thresh                = 0.01,
+    m.opts                  = my.opts,
+    num.parallel.samples    = 1,
+    mc.cores.per.sample     = 30,
+    seed                    = 2351,
+    use.sparse.assign       = TRUE,
+    drop.low.mut.samples    = FALSE)
+  
+  expect_equal(nrow(retval4$proposed.assignment), 2)
   
   sapply(output.dir, FUN = unlink, recursive = TRUE)
 })