+++ NEWS.md
@@ -4,6 +4,10 @@
 
 * Added new internal function `NumFromId` and removed `ICAMSxtra` from the dependency packages.
 
+* Enabled argument `target.sig.index` to also take the signature id as input in
+function `SignaturePresenceTest`. Changed argument `m.opts` to use return value
+from `DefaultManyOpts` by default. Changed the default value for argument `mc.cores` to 2.
+
 # 2.1.2
 * Enabled argument `target.sig.index` to also take the signature id as input in
 function `SignaturePresenceTest1`. Changed argument `m.opts` to use return value